+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This is a changelog for Piwik platform developers. All changes for our HTTP API'
 * As we are now using CLDR formats for all languages, some time formats were even changed in english. Attributes like `prettyDate` in API responses might so have been changed slightly.
 * The following deprecated classes have been removed:
     * `Piwik\Registry`: use the container instead
+    * `Piwik\TaskScheduler`: use `Piwik\Scheduler\Scheduler` instead
+    * `Piwik\ScheduledTask`: use `Piwik\Scheduler\Task` instead
 * The following deprecated methods have been removed:
     * `Piwik\Common::getContinentsList()`: use `Piwik\Intl\Data\Provider\RegionDataProvider::getContinentList()` instead
     * `Piwik\Common::getCountriesList()`: use `Piwik\Intl\Data\Provider\RegionDataProvider::getCountryList()` instead